move Auth functions into User model
