+++ src/popup/popup.js
@@ -1,7 +1,7 @@
 import Options from './options/options.js'
 import PopupDOM from './popup.dom.js'
 import style from './popup.css'
-import * as Auth from '../model/Auth.js'
+import * as User from '../model/User.js'
 import * as Comment from '../model/Comment.js'
 import * as Group from '../model/Group.js'
 
@@ -10,11 +10,11 @@ function initApp() {
   let DOM = new PopupDOM()
 
   /**
-   * closeLoginPrompt and loginPrompt are passed to Auth.js,
+   * closeLoginPrompt and loginPrompt are passed to User.js,
    *   and are called when the login/logout status of the 
    *   user changes.
    */
-  Auth.start(closeLoginPrompt, loginPrompt)
+  User.initializeFirebase(closeLoginPrompt, loginPrompt)
 
   DOM.submitBtn.addEventListener('click', submit)
   DOM.optionsBtn.addEventListener('click', openOptions)
@@ -32,7 +32,7 @@ function initApp() {
   }
 
   function logout() {
-    Auth.logout()
+    User.logout()
   }
 
   function inputKeydown(e) {
@@ -43,12 +43,12 @@ function initApp() {
   }
 
   function loginPrompt() {
-    if (!Auth.isUserLoggedIn()) {
+    if (!User.isUserLoggedIn()) {
       let el = document.createElement('div')
       el.innerHTML = DOM.getLoginOverlay();
       document.body.appendChild(el)
       setTimeout(() => document.getElementsByClassName('login-overlay')[0].classList.remove('loading'), 200)
-      document.getElementById('welcome-login-btn').addEventListener('click', () => Auth.login())
+      document.getElementById('welcome-login-btn').addEventListener('click', () => User.login())
     }
   }
